Delete and Update redirect to main page. The data rerenders kinda buggy

+++ src/components/ContactDetails/index.jsx
@@ -1,12 +1,13 @@
 import { useState, useEffect } from 'react'
-import { useParams, useNavigate } from 'react-router-dom'
+import { useParams, useNavigate, Routes, Route } from 'react-router-dom'
+import UpdateContact from '../UpdateContact'
 
 import './style.css'
 
 function ContactDetails(props) {
     const [contact, setContact] = useState(null)
     const { id } = useParams()
-    const { contacts } = props
+    const { contacts, toggle, setToggle } = props
     const navigate = useNavigate()
 
     useEffect(() => {
@@ -20,18 +21,24 @@ function ContactDetails(props) {
     if(!contact) return <p>Loading contact...</p>
 
     const handleClick = (event) => {
-        fetch(`https://boolean-api-server.fly.dev/nora-hansen/contact/${contact.id}`,
-        {
-            method: "DELETE",
-        })
-        navigate("/contactlist")
+        if(event.target.name === "delete") {
+            fetch(`https://boolean-api-server.fly.dev/nora-hansen/contact/${contact.id}`,
+            {
+                method: "DELETE",
+            })
+            setToggle(toggle => !toggle)
+            navigate("/contactlist")
+        }   else if (event.target.name === "update") {
+            navigate(`/contact/${id}/edit`)
+        }
     }
 
     return(
         <div className="contact-details">
             <h2>{contact.firstName} {contact.lastName}</h2>
             <p>{contact.street} {contact.city}</p>
-            <button onClick={handleClick}>Delete</button>
+            <button name="delete" onClick={handleClick}>Delete</button>
+            <button name="update" onClick={handleClick}>Edit</button>
         </div>
     )
 }
